Rename modules + placeholders
